refactor(web): SSE 驱动 gameState、扁平化错误处理、graph-canvas 抽离
- 移除 submitInteraction 后冗余的 getSessionSnapshot 拉取(submit-form 不改 character/schema，character.upserted 经 SSE 全覆盖)
- submitInteraction 3 层 try-catch 扁平化，严格保持降级行为
- graph-canvas 抽 graph-canvas-sync.ts 纯函数 + useReducer，渲染行为不变
- web 267 测试全过
